+++ ui/jobs/jobs.js
@@ -6,7 +6,7 @@ import { Util } from '../../resources/common.js';
 import ZoneInfo from '../../resources/zone_info.js';
 import ZoneId from '../../resources/zone_id.js';
 import { kWellFedContentTypes, kLevelMod, kMPCombatRate, kMPNormalRate, kMPUI1Rate, kMPUI2Rate, kMPUI3Rate, kMPTickInterval } from './constants.js';
-import { BuffTracker } from './BuffTracker.js';
+import { BuffTracker } from './buff_tracker.js';
 import { RegexesHolder, computeBackgroundColorFrom, doesJobNeedMPBar, setupComboTracker, makeAuraTimerIcon } from './utils.js';
 
 import {
